ipa: neo: camhelper: ox03c10: Fix min analog gain value
The function controlInfoMapGetAnalogGainRange() is providing analog gain range
for which the AGC algorithm should operate.
The minimal analog gain value for long capture depends on HCG and LCG ratio
constraints.
The ratio between HCG and LCG min values should be the maximum between:
- the gain conversion ratio HCG/LCG (read from OTP with a default to 7.32)
- the ratio between total exposure of Long and Short captures (currently set to 32)

This patch fixes the minimal analog gain value where the gain conversion
ratio (typically 7.32) was applied, resulting in a lower min gain value used by
the AGC algorithm (expected min gain value is 32).

The impact for this improper min gain value was observed when total exposure time
was lower than max exposure time (32498.01us).
In this case, the AGC algorithm was distributing the total exposure time with an
analog gain lower than 32.
This impact was not visible since the AGC distribution post processing the AGC
algorithm outputs (either from CameraHelper or from the ox03c10 kernel driver)
was rectifying this distribution.

Note that the uGuzzi IPA is not using the controlInfoMapGetAnalogGainRange() and
controlInfoMapGetExposureRange(), it is retrieving the value range from DTP instead.
Hence this effect was not reproduced with the uGuzzi IPA. However the CameraHelper
implementation should be aligned.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>