scripts: gen-version: Fix dirty tree detection with out-of-source build
Neo-ipa-uguzzi, when built for the Yocto distribution, has a spurious
`-dirty` suffix in the git version associated to the library.

Git version and dirty tree detection are generated at compilation time
by the script gen-version.h, that eventually appends the `-dirty` suffix.
Dirty detection is done using the `git diff-index` command that requires
a prior `git update-index --refresh` to avoid a false positive.
However the index refresh is currently done only for in-source build.

Yocto builder (bitbake) uses out-of-source build. In that context, the
false positives come from file timestamp change and hard link creation
in the source tree, done by some of the builder tasks.

This changes is to have `git update-index --refresh` command executed
also in case of out-of-source build, to fix the false dirty
detections.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>