ipa: neo: Calculate line duration based on sensor hblank
Until now, the line duration was calculated based on the minimum
line length. However, it may be possible to request fps control by
adjusting the sensor hblank and vblank. Hence the line duration is now
calculated based on the sensor hblank.

For this, the hblank and vblank are passed to the CameraHelper by adding
those parameters part of the struct CameraMode.
And a new CameraHelper::hblankToLineLength(uint32_t hblank) is defined to calculate
the line duration.

This patch also changes the type of minLineLength/maxLineLength from
struct CameraMode to Duration instead of pixel length. It is to be aligned
with the type used from the RPi struct CameraMode.

It also adds CameraHelper::exposure() and CameraHelper::exposureLines()
to convert between line length and line duration.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>