utils: Add stream mode trigger for the selected tuning configuration
For now, the tuning info to configure is selected based on the sensor
entity or model, the resolution and the bit depth.
This patch adds the sensor stream mode as a trigger selection for the
tuning info.
Indeed the tuning info (DTP file or tuning parameters) can be different
depending on the sensor stream mode (such as HDR stream mode).

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>