ipa: neo: mx95mbcam: Add checksum verification on embedded data
Sensor ox03c10 appends a CRC-32C checksum at the end of the embedded
data as a way to check its integrity.
This change adds support to the MdParserOmniOx class, with a new
argument passed to the constructor to handle an optional CRC on the
embedded data content.
Implementation of CRC-32C (Castagnoli) is provided to support the
ox03a10 CRC.

Parsing error messages are demoted from error to warning as they are
not fatal for the system since DelayedControl values take over in case
of embedded data unavailability or corruption.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>