data: Rename the os08a20 DTP as database_os08a20_SDR_3840-2160_12bpp.bin
The actual os08a20 DTP is tuned for single exposure, 4k resolution (3840x2160)
and 12 bit depth. It is renamed explicitly according to the resolution, bit depth
and as SDR for Single Dynamic Range. This will help distinguish with other DTP
supporting other resolution or other bit depth or for HDR mode.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>