data: Define the socket port for the sensor model
This patch changes the socket port to be defined for
the sensor model instead of the sensor entity.
This helps to support any sensor entity for the same model
when connecting with Tuning Tool.
Indeed, when using different board, such as the i.MX95 FRDM board,
the entity of a supported sensor model can differ (different i2c bus).

With this change, the socket port is the same for all the entities
of the same model. This is not impacting since connection
with Tuning Tool only supports one sensor at a time.

It is still possible to customize this configuration file with
socket port to be defined for a sensor entity.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>