ipa: neo: ox05b1s: Add multi-capture controls used in RGBIr dual mode
In RGBIr dual mode, the ox05b1s is using the multi-capture controls
V4L2_CID_AGAIN_MULTI and V4L2_CID_EXPOSURE_MULTI instead of the
single-capture controls.
The convention is to provide the:
- RGB context values to the span position 0 of the control
- Ir context values to the span position 1 of the control

Hence the values stored in the sensor metadata control list are
adapted accordingly and stored the same way as the sensor control list.

A new CameraHelper enum type SensorContextTypes is created to
map with the IPAContextType enums.
The argument SensorContextTypes is added to controlListSetAGC()
to provide the frame context (RGB or Ir).

In RGBIr dual mode, the controls are sent for the context Ir only,
Indeed, after processing the Ir context, the multi controls are updated
with both context of the frame.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>