ipa: neo: Add call to uguzzi_config() to get initial settings
To get the initial sensor and ISP settings, the uguzzi_process()
should be called with NULL pointers for input sensor data and statistics.
However it is also required to call first uguzzi_config() with any valid
config to make sure that the ON_CONFIG settings will be also reported.
For this, the config with the CMD_AE_MODE applies.

This is solving issue where the ON_CONFIG settings were not applied in a
config/start/stop/start calls sequence.
The ISP STAT and RGBIR block settings (configured as ON_CONFIG) were not
set due to no uguzzi_config() applied between the stop() and the start()
calls.
This issue was generating logs warning with histogram mismatch and with
empty histogram sums.
This was reproduced when launching lc-compliance tests triggering such
calls sequence.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>